+++ crates/gmsol/src/alt.rs
@@ -92,7 +92,7 @@ impl<C: Deref<Target = impl Signer> + Clone> AddressLookupTableOps<C> for crate:
         let rpc = self
             .store_transaction()
             .program(address_lookup_table::program::ID)
-            .pre_instruction(ix);
+            .pre_instruction(ix, false);
 
         Ok((rpc, address))
     }
@@ -117,7 +117,7 @@ impl<C: Deref<Target = impl Signer> + Clone> AddressLookupTableOps<C> for crate:
             let rpc = self
                 .store_transaction()
                 .program(address_lookup_table::program::ID)
-                .pre_instruction(ix);
+                .pre_instruction(ix, false);
             tx.try_push(rpc)?;
         }
         Ok(tx)
@@ -128,7 +128,7 @@ impl<C: Deref<Target = impl Signer> + Clone> AddressLookupTableOps<C> for crate:
         let ix = address_lookup_table::instruction::deactivate_lookup_table(*alt, payer);
         self.store_transaction()
             .program(address_lookup_table::program::ID)
-            .pre_instruction(ix)
+            .pre_instruction(ix, false)
     }
 
     fn close_alt(&self, alt: &Pubkey) -> TransactionBuilder<C> {
@@ -136,6 +136,6 @@ impl<C: Deref<Target = impl Signer> + Clone> AddressLookupTableOps<C> for crate:
         let ix = address_lookup_table::instruction::close_lookup_table(*alt, payer, payer);
         self.store_transaction()
             .program(address_lookup_table::program::ID)
-            .pre_instruction(ix)
+            .pre_instruction(ix, false)
     }
 }