+++ crates/gmsol/src/client.rs
@@ -1224,7 +1224,7 @@ impl<C: Clone + Deref<Target = impl Signer>> SystemProgramOps<C> for Client<C> {
         }
         Ok(self
             .store_transaction()
-            .pre_instruction(transfer(&self.payer(), to, lamports)))
+            .pre_instruction(transfer(&self.payer(), to, lamports), false))
     }
 }