+++ crates/sdk/src/client/ops/oracle.rs
@@ -69,7 +69,7 @@ impl<C: Deref<Target = impl Signer> + Clone> OracleOps<C> for crate::Client<C> {
 
         let builder = self
             .store_transaction()
-            .pre_instruction(create)
+            .pre_instruction(create, false)
             .anchor_accounts(accounts::InitializeOracle {
                 payer,
                 authority: authority.copied().unwrap_or(payer),